requirements.txt
pydub==0.25.1
SpeechRecognition==3.10.1
vosk==0.3.44
python-telegram-bot==20.8
python-dotenv==1.0.0
